Add styles to navbar
Wdit project tree
