+++ src/components/App.css
@@ -1,3 +1,13 @@
 .App {
-  text-align: center;
+  display: flex;
+  flex-direction: column;
+  width: 1400px;
+  height: 1097px;
+  margin: 0 40px 65px 0;
+  padding: 0 0 794px;
+  border-radius: 4px;
+  box-shadow: 0 0 60px 0 rgba(0, 0, 0, 0.05);
+  background-color: #fafafa
+
+;
 }
