+++ src/components/App.js
@@ -1,6 +1,6 @@
 import './App.css';
-import BooksList from '../containers/BooksList';
-import BooksForm from '../containers/BooksForm';
+import BooksList from '../containers/BooksList/BooksList';
+import BooksForm from '../containers/BooksForm/BooksForm';
 
 const categories = ['Action', 'Biography', 'History', 'Horror', 'Kids', 'Learning', 'Sci-Fi'];
